Add files via upload
The efficient frontier graphically depicts the tremendous benefit of portfolio diversification. The curvature of the line is integral in revealing how diversification improves a portfolio's risk / reward profile. It also reveals that there is a diminishing marginal return to risk. In other words, adding more risk to a portfolio does not gain an equal amount of return. Optimal portfolios that comprise the efficient frontier tend to have a higher degree of diversification than the sub-optimal ones, which are typically less diversified.